+++ src/main/java/com/qa/Consumer/persistence/repository/MongoCohortRepo.java
@@ -1,9 +1,9 @@
-package com.qa.InspectorFacade.persistence.repository;
+package com.qa.Consumer.persistence.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.qa.InspectorFacade.persistence.domain.SentCohort;
+import com.qa.Consumer.persistence.domain.SentCohort;
 
 @Repository
 public interface MongoCohortRepo extends MongoRepository<SentCohort, Long>{
