+++ src/main/java/com/qa/Consumer/persistence/repository/MongoTraineeRepo.java
@@ -1,9 +1,9 @@
-package com.qa.InspectorFacade.persistence.repository;
+package com.qa.Consumer.persistence.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.qa.InspectorFacade.persistence.domain.SentTrainee;
+import com.qa.Consumer.persistence.domain.SentTrainee;
 
 @Repository
 public interface MongoTraineeRepo extends MongoRepository<SentTrainee, Long> {
