+++ src/main/java/com/qa/Consumer/receiver/TraineeReceiver.java
@@ -1,11 +1,11 @@
-package com.qa.InspectorFacade.receiver;
+package com.qa.Consumer.receiver;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jms.annotation.JmsListener;
 import org.springframework.stereotype.Component;
 
-import com.qa.InspectorFacade.persistence.domain.SentTrainee;
-import com.qa.InspectorFacade.persistence.repository.MongoTraineeRepo;
+import com.qa.Consumer.persistence.domain.SentTrainee;
+import com.qa.Consumer.persistence.repository.MongoTraineeRepo;
 
 @Component
 public class TraineeReceiver {
