+++ src/main/java/com/qa/Consumer/receiver/TrainerReceiver.java
@@ -1,11 +1,11 @@
-package com.qa.InspectorFacade.receiver;
+package com.qa.Consumer.receiver;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jms.annotation.JmsListener;
 import org.springframework.stereotype.Component;
 
-import com.qa.InspectorFacade.persistence.domain.SentTrainer;
-import com.qa.InspectorFacade.persistence.repository.MongoTrainerRepo;
+import com.qa.Consumer.persistence.domain.SentTrainer;
+import com.qa.Consumer.persistence.repository.MongoTrainerRepo;
 
 @Component
 public class TrainerReceiver {
